Refactored Bottom Navigation
This is now a distinct component.
