+++ src/components/shared/NavigationBackground.jsx
@@ -6,7 +6,15 @@ import {baseColors} from '../../styles/base/colors.js'
 
 function NavigationBackground(props) {
   return (
-    <div className={css(styles.backgroundStyles)}>
+    <div
+      className={
+        css(
+          styles.backgroundStyles,
+          (props.shadowVariant === 'up') && styles.upShadow,
+          (props.shadowVariant === 'down') && styles.downShadow
+        )
+      }
+    >
       <SpacingInset size={'m'} >
         {props.children}
       </SpacingInset>
@@ -17,6 +25,12 @@ function NavigationBackground(props) {
 const styles = StyleSheet.create({
   backgroundStyles: {
     backgroundColor: baseColors.ultraLightGrey
+  },
+  downShadow: {
+    boxShadow: '0px 2px 4px 1px rgba(0,0,0,0.24)'
+  },
+  upShadow: {
+    boxShadow: '0px -2px 4px 1px rgba(0,0,0,0.24)'
   }
 })
